Reorganize file names by type
