+++ editor/filter_attribute_append_test.go
