+++ editor/filter_attribute_remove.go
