+++ editor/filter_attribute_remove_test.go
