+++ editor/filter_attribute_set.go
