+++ editor/filter_block_append_test.go
