+++ editor/filter_block_remove.go
