+++ editor/filter_block_remove_test.go
