+++ editor/operator_edit.go
