+++ editor/sink_attribute_get.go
