+++ editor/sink_attribute_get_test.go
